+++ src/react/admin.js
@@ -10,8 +10,18 @@ import { IntlProvider, addLocaleData } from 'react-intl'
 import enLocaleData from 'react-intl/locale-data/en'
 import deLocaleData from 'react-intl/locale-data/de'
 
+// Store
+import { Provider } from 'react-redux'
+import { applyMiddleware, createStore } from 'redux'
+import { composeWithDevTools } from 'redux-devtools-extension'
+import thunkMiddleware from 'redux-thunk'
+import reducers from './reducers'
+
 // Components
 import App from './components/app'
+
+// Redux Store
+const store = createStore(reducers, composeWithDevTools(applyMiddleware(thunkMiddleware)))
 
 // logging and debugging
 const debug = require('debug')('Admin')
@@ -40,9 +50,11 @@ const renderApp = async () => {
     const result = await axios.get(`${wpGlobals.asset_path}/languages/${lang}.json`)
 
     render(
-      <IntlProvider locale={lang} messages={result.data}>
-        <App wpGlobals={wpGlobals} />
-      </IntlProvider>,
+      <Provider store={store}>
+        <IntlProvider locale={lang} messages={result.data}>
+          <App wpGlobals={wpGlobals} />
+        </IntlProvider>
+      </Provider>,
       document.getElementById('plugin-name-app'),
     )
   } catch (err) {