+++ src/react/components/app/__tests__/component.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import App from './'
-import { createComponentWithIntl } from '../../../../jest/intl-enzyme-setup'
+import { createComponentWithIntl } from '../../../../../jest/intl-enzyme-setup'
+import App from '../component'
 
 // useful tipps for testing with JEST
 // - https://marmelab.com/blog/2015/06/24/jest-in-practice.html