+++ src/react/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import { reducer as app } from './app'
+
+export default combineReducers({
+  app,
+})