+++ src/react/utils/__mocks__/api-client.js
@@ -1,0 +1,3 @@
+export default {
+  request: jest.fn().mockResolvedValue({ data: { data: { prop: 'some-data' } } }),
+}